fixed #1

+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main {
         int eaters = 5; // сколько людей будут есть
         int water = 3000; // миллилитров воды
         int potatoes = 5; // картофелин
-        int chicken = 6; // куриных бёдер
+        double chicken = 6; // куриных бёдер
         int spices = 10; // ложек специй
 
         System.out.println("Сварили суп. На одного человека вышло:");
